Change webvr initialization to check for available devices before creating the VR buttons
A recent change to the WebVR spec forces the HMD.requestPresent call to be done inside a user gesture (ie a mouse click handler),
so we need to poll the devices beforehand. Otherwise when the getVRDisplays promise returns we're no longer inside the user gesture
started by clicking on the button and the call to requestPresent will fail.